feat(ups-002): implementacion 2 direcciones del laberinto

+++ src/EjerciciosPD.java
@@ -3,19 +3,19 @@ import java.util.Map;
 
 public class EjerciciosPD {
 
-    public int getFibonacci(int n){
+    public long getFibonacci(int n){
         if(n <=1){
             return n;
         }
         return getFibonacci(n-1) + getFibonacci(n-2);
     }
 
-    public int getFibonacciPD(int n){
-        Map<Integer,Integer> cache = new HashMap<>();
+    public long getFibonacciPD(int n){
+        Map<Integer,Long> cache = new HashMap<>();
         return getFibonacciPDHelper(n, cache);
     }
 
-    private int getFibonacciPDHelper(int n, Map<Integer,Integer> cache){
+    private long getFibonacciPDHelper(int n, Map<Integer,Long> cache){
         if(n<=1){
             return n;
         }
@@ -24,7 +24,7 @@ public class EjerciciosPD {
             return cache.get(n);
         }
 
-        int resultado = getFibonacciPDHelper(n-1, cache) + getFibonacciPDHelper(n-2, cache);
+        long resultado = getFibonacciPDHelper(n-1, cache) + getFibonacciPDHelper(n-2, cache);
         cache.put(n,resultado);
 
         return resultado;
